requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.1
bootstrap4==0.1.0
captcha==0.3
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.5
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-extensions==3.0.9
django-json-field==0.5.7
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-ranged-response==0.2.0
django-resized==0.3.11
django-simple-captcha==0.5.13
djangorestframework==3.12.1
easy-thumbnails==2.7.1
filelock==3.0.12
freeze==3.0
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
install==1.3.4
jsonfield==3.1.0
packaging==20.4
phonenumbers==8.12.13
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
Pygments==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
urllib3==1.25.11
virtualenv==20.2.1
